Put PWM register map in separate package with other constants for students to see how to write a clean testbench.
Fix small bug in testbench where the PWM output can be stuck at '1' when we send a STOP command. We now force the output to '0' at the end of the ongoing period when the PWM is instructed to STOP.

Rename "started" to "run" in pwm.vhd.

Make sure readdata is 0 when reading the write-only control register.